Fix: changes on the desing of the component
